+++ test/widget_test.dart
@@ -3,35 +3,44 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:registrar/registrar.dart';
 import 'package:mvvm_plus/mvvm_plus.dart';
 
-import 'unit_test.dart';
-
 const _number = 42;
+const _floatDefault = 3.14159;
+const _floatUpdated = 42.42;
+const _stringDefault = 'Default';
+const _stringUpdated = 'Updated';
+const _registeredStringDefault = 'Registered Default';
+const _registeredStringUpdated = 'Registered Updated';
+const _namedStringDefault = 'Named Default';
+const _namedStringUpdated = 'Named Updated';
+const _propertyName = 'Property Name';
 const _viewModelName = 'ViewModel Name';
 
 /// Test app for all widget tests
 ///
-/// [listen] true to listen to [MyNotifier]
-/// [register] true to register [MyWidgetViewModel]
-/// [name] is option name of registered [MyWidgetViewModel]
+/// [listenToRegistrar] true to listen to [My]
+/// [registerViewModel] true to register [MyTestWidgetViewModel]
+/// [viewModelName] is option name of registered [MyTestWidgetViewModel]
 Widget testApp({
-  required bool listen,
-  required bool register,
-  required String? name,
+  required bool listenToRegistrar,
+  required bool registerViewModel,
+  required String? viewModelName,
 }) =>
     MaterialApp(
       home: Registrar(
-        builder: () => MyNotifier(),
-        child: MyWidget(
-          listen: listen,
-          register: register,
-          name: name,
+        builder: () => MyModel(),
+        child: MyTestWidget(
+          listenToRegistrar: listenToRegistrar,
+          registerViewModel: registerViewModel,
+          viewModelName: viewModelName,
         ),
       ),
     );
 
 /// The [Registrar] service
-class MyNotifier extends ChangeNotifier {
+class MyModel extends Model {
   int number = _number;
+
+  final myFloatProperty = Property<double>(_floatDefault);
 
   void incrementNumber() {
     number++;
@@ -40,61 +49,67 @@ class MyNotifier extends ChangeNotifier {
 }
 
 /// The [View]
-class MyWidget extends View<MyWidgetViewModel> {
-  MyWidget({
+class MyTestWidget extends View<MyTestWidgetViewModel> {
+  MyTestWidget({
     super.key,
-    required bool listen,
-    required bool register,
-    required String? name,
+    required bool listenToRegistrar,
+    required bool registerViewModel,
+    required String? viewModelName,
   }) : super(
-      viewModelBuilder: () =>
-          MyWidgetViewModel(
-            listen: listen,
-            register: register,
-            name: name,
-          ));
+            viewModelBuilder: () => MyTestWidgetViewModel(
+                  listenToRegistrar: listenToRegistrar,
+                  register: registerViewModel,
+                  name: viewModelName,
+                ));
 
   @override
   Widget build(BuildContext _) {
+    final float = listenTo<Property<double>>(notifier: get<MyModel>().myFloatProperty).value;
     return Column(
       children: [
         Text('${viewModel.number}'),
+        Text(viewModel.myStringProperty.value),
+        Text(viewModel.myRegisteredStringProperty.value),
+        Text(viewModel.myNamedStringProperty.value),
+        Text('$float'),
       ],
     );
   }
 }
 
 /// The [ViewModel]
-class MyWidgetViewModel extends ViewModel {
-  MyWidgetViewModel({
-    this.listen = false,
+class MyTestWidgetViewModel extends ViewModel {
+  MyTestWidgetViewModel({
+    this.listenToRegistrar = false,
     super.register,
     super.name,
   });
 
-  final bool listen;
-  late final MyNotifier myNotifier;
+  final bool listenToRegistrar;
+  late final MyModel myModel;
+  late final myStringProperty = Property<String>(_stringDefault)..addListener(buildView);
+  late final myRegisteredStringProperty = Property<String>(_registeredStringDefault)..addListener(buildView);
+  late final myNamedStringProperty = Property<String>(_namedStringDefault)..addListener(buildView);
 
   @override
   void initState() {
     super.initState();
-    if (listen) {
+    if (listenToRegistrar) {
       // listen twice so can later test that only one listener added
-      listenTo<MyNotifier>(); // 1st listen
-      myNotifier = listenTo<MyNotifier>(); // 2nd listen
+      listenTo<MyModel>(); // 1st listen
+      myModel = listenTo<MyModel>(); // 2nd listen
     } else {
-      myNotifier = get<MyNotifier>();
+      myModel = get<MyModel>();
     }
   }
 
-  int get number => myNotifier.number;
+  int get number => myModel.number;
 }
 
 /// Test app for widget subclassed from [ViewWithStatelessViewModel]
-Widget statelessTestApp({required bool listen}) =>
-    MaterialApp(
+Widget statelessTestApp({required bool listen}) => MaterialApp(
       home: Registrar(
-        builder: () => MyNotifier(),
+        builder: () => MyModel(),
         child: MyStatelessView(listen: listen),
       ),
     );
@@ -110,75 +125,92 @@ class MyStatelessView extends ViewWithStatelessViewModel {
 
   @override
   Widget build(BuildContext context) {
-    return listen ? Text('${listenTo<MyNotifier>().number}') : Text('${get<MyNotifier>().number}');
+    return listen ? Text('${listenTo<MyModel>().number}') : Text('${get<MyModel>().number}');
   }
 }
 
 void main() {
   setUp(() {
     /// Ensure no residuals
-    expect(Registrar.isRegistered<MyNotifier>(), false);
-    expect(Registrar.isRegistered<MyWidgetViewModel>(), false);
+    expect(Registrar.isRegistered<MyModel>(), false);
+    expect(Registrar.isRegistered<MyTestWidgetViewModel>(), false);
   });
 
   tearDown(() {
     /// Ensure no residuals
-    expect(Registrar.isRegistered<MyNotifier>(), false);
-    expect(Registrar.isRegistered<MyWidgetViewModel>(), false);
+    expect(Registrar.isRegistered<MyModel>(), false);
+    expect(Registrar.isRegistered<MyTestWidgetViewModel>(), false);
   });
 
-  group('MyWidget', () {
-    testWidgets('not listening, registered, or named ViewModel does not update value', (WidgetTester tester) async {
-      await tester.pumpWidget(testApp(listen: false, register: false, name: null));
+  group('MyTestWidget', () {
+    testWidgets('not listening to Registrar, not registered, and not named ViewModel does not update value',
+        (WidgetTester tester) async {
+      await tester.pumpWidget(testApp(listenToRegistrar: false, registerViewModel: false, viewModelName: null));
 
-      expect(Registrar.isRegistered<MyNotifier>(), true);
+      expect(Registrar.isRegistered<MyModel>(), true);
       expect(find.text('$_number'), findsOneWidget);
 
-      Registrar.get<MyNotifier>().incrementNumber();
+      Registrar.get<MyModel>().incrementNumber();
       await tester.pump();
 
       // expect does not increment b/c not listening
       expect(find.text('$_number'), findsOneWidget);
     });
 
-    testWidgets('listening but not registered ViewModel shows correct values', (WidgetTester tester) async {
-      await tester.pumpWidget(testApp(listen: true, register: false, name: null));
+    testWidgets('listening to Registrar but not registered ViewModel shows correct values',
+        (WidgetTester tester) async {
+      await tester.pumpWidget(testApp(listenToRegistrar: true, registerViewModel: false, viewModelName: null));
 
-      expect(Registrar.isRegistered<MyNotifier>(), true);
+      expect(Registrar.isRegistered<MyModel>(), true);
       expect(find.text('$_number'), findsOneWidget);
 
-      Registrar.get<MyNotifier>().incrementNumber();
+      Registrar.get<MyModel>().incrementNumber();
       await tester.pump();
 
       expect(find.text('${_number + 1}'), findsOneWidget);
     });
 
-    testWidgets('listening and registered but not named ViewModel shows correct values', (WidgetTester tester) async {
-      await tester.pumpWidget(testApp(listen: true, register: true, name: null));
+    testWidgets('listening to Registrar and registered ViewModel  but not named ViewModel shows correct values',
+        (WidgetTester tester) async {
+      await tester.pumpWidget(testApp(listenToRegistrar: true, registerViewModel: true, viewModelName: null));
 
+      expect(Registrar.isRegistered<MyTestWidgetViewModel>(), true);
+      expect(Registrar.get<MyTestWidgetViewModel>().number, _number);
+
+      // expect default values
       expect(find.text('$_number'), findsOneWidget);
-      expect(Registrar.isRegistered<MyWidgetViewModel>(), true);
-      expect(Registrar
-          .get<MyWidgetViewModel>()
-          .number, _number);
+      expect(find.text(_stringDefault), findsOneWidget);
+      expect(find.text(_registeredStringDefault), findsOneWidget);
+      expect(find.text(_namedStringDefault), findsOneWidget);
+      expect(find.text('$_floatDefault'), findsOneWidget);
 
-      Registrar.get<MyNotifier>().incrementNumber();
+      // change values
+      Registrar.get<MyModel>().incrementNumber();
+      Registrar.get<MyTestWidgetViewModel>().myStringProperty.value = _stringUpdated;
+      Registrar.get<MyTestWidgetViewModel>().myRegisteredStringProperty.value = _registeredStringUpdated;
+      Registrar.get<MyTestWidgetViewModel>().myNamedStringProperty.value = _namedStringUpdated;
+      Registrar.get<MyModel>().myFloatProperty.value = _floatUpdated;
+
       await tester.pump();
 
+      // expect updated values
       expect(find.text('${_number + 1}'), findsOneWidget);
+      expect(find.text(_stringUpdated), findsOneWidget);
+      expect(find.text(_registeredStringUpdated), findsOneWidget);
+      expect(find.text(_namedStringUpdated), findsOneWidget);
+      expect(find.text('$_floatUpdated'), findsOneWidget);
     });
 
-    testWidgets('listening, registered, and named ViewModel shows correct values', (WidgetTester tester) async {
-      await tester.pumpWidget(testApp(listen: true, register: true, name: _viewModelName));
+    testWidgets('listening to Registrar, registered and named ViewModel shows correct values',
+        (WidgetTester tester) async {
+      await tester.pumpWidget(testApp(listenToRegistrar: true, registerViewModel: true, viewModelName: _viewModelName));
 
       expect(find.text('$_number'), findsOneWidget);
-      expect(Registrar.isRegistered<MyWidgetViewModel>(), false);
-      expect(Registrar.isRegistered<MyWidgetViewModel>(name: _viewModelName), true);
-      expect(Registrar
-          .get<MyWidgetViewModel>(name: _viewModelName)
-          .number, _number);
+      expect(Registrar.isRegistered<MyTestWidgetViewModel>(), false);
+      expect(Registrar.isRegistered<MyTestWidgetViewModel>(name: _viewModelName), true);
+      expect(Registrar.get<MyTestWidgetViewModel>(name: _viewModelName).number, _number);
 
-      Registrar.get<MyNotifier>().incrementNumber();
+      Registrar.get<MyModel>().incrementNumber();
       await tester.pump();
 
       expect(find.text('${_number + 1}'), findsOneWidget);
@@ -189,10 +221,10 @@ void main() {
     testWidgets('non-listening stateless View does not update', (WidgetTester tester) async {
       await tester.pumpWidget(statelessTestApp(listen: false));
 
-      expect(Registrar.isRegistered<MyNotifier>(), true);
+      expect(Registrar.isRegistered<MyModel>(), true);
       expect(find.text('$_number'), findsOneWidget);
 
-      Registrar.get<MyNotifier>().incrementNumber();
+      Registrar.get<MyModel>().incrementNumber();
       await tester.pump();
 
       // expect number did not increment (because not listening)
@@ -202,10 +234,10 @@ void main() {
     testWidgets('listening stateless View updates', (WidgetTester tester) async {
       await tester.pumpWidget(statelessTestApp(listen: true));
 
-      expect(Registrar.isRegistered<MyNotifier>(), true);
+      expect(Registrar.isRegistered<MyModel>(), true);
       expect(find.text('$_number'), findsOneWidget);
 
-      Registrar.get<MyNotifier>().incrementNumber();
+      Registrar.get<MyModel>().incrementNumber();
       await tester.pump();
 
       expect(find.text('${_number + 1}'), findsOneWidget);