+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.0
+
+Added Model.
+Added "notifier" parameter to listenTo to support listening to ValueNotifiers.
+
 ## 0.3.1
 
 Corrected changelog.